+++ web-app/src/components/tournament/details/ScoreInputDialog.vue
